app/pyproject.toml
[project]
name = "care-coordinator-assistant-app"
version = "0.1.0"
description = "Care Coordinator Assistant App"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "langgraph>=0.6.5",
    "pydantic-settings>=2.10.1",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "httpx>=0.27.2",
    "ipython>=9.4.0",
    "uvicorn>=0.35.0",
    "python-levenshtein>=0.27.1",
]

[dependency-groups]
test = [
    "pytest>=8.4.1",
]

[tool.uv.workspace]
members = [
    "client",
]
